python/requirements.txt
# pyannote.audio 声纹识别依赖

# 核心库
pyannote.audio==3.1.1
torch>=2.0.0
torchaudio>=2.0.0

# 音频处理
librosa>=0.10.0
soundfile>=0.12.1
scipy>=1.10.0

# 工具
numpy>=1.24.0
pandas>=2.0.0

# 如果需要使用HuggingFace模型
# huggingface_hub>=0.16.0
